[MOD] moved popup alert creation content into a component, fixed title of graph space, fixed link to individual alert from homepage
