+++ src/Components/Alerts/showAlerts.js
@@ -25,7 +25,7 @@ export default function ShowAlerts ({chosenAlert}) {
               <div>
                 {chosenAlert?.file
                   ? <button id='alert-btn'>
-                    <a style={{ textDecoration: 'none', color: 'white' }} href={chosenAlert?.file ? chosenAlert.file : ''} target='_blank' rel='noopener noreferrer'>
+                    <a style={{ textDecoration: 'none', color: 'white' }} href={chosenAlert.file} target='_blank' rel='noopener noreferrer'>
                       Télécharger la pièce-jointe
                     </a>
                   </button>
